+++ CHANGELOG.md
@@ -3,8 +3,8 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
- 
-## [Unreleased]
+
+## Unreleased
 ---
 ## Unreleased - [2.0.0]
 ### Changed
@@ -30,6 +30,14 @@ Object has different properties!
 - Ability to remove message listeners.
 
 ---
+
+## [1.2.1] - 2021-06-03
+### Added:
+- Logging for heartbeat messages.
+ 
+## [1.2.0] - 2021-05-29
+### Added:
+- BACKPORT [2.0.0]: Support for History-Info header.
 
 ## [1.1.0] - 2021-05-07
 ### Changed